backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
email-validator
requests
pydantic
bcrypt==4.0.1
python-dotenv
